+++ tests/destructors.rs
@@ -14,15 +14,16 @@ fn resource_destructor_request() {
     let destructor_called_global = destructor_called.clone();
 
     let mut server = TestServer::new();
-    server
-        .display
-        .create_global::<ServerOutput, _>(3, move |newo, _, _| {
+    server.display.create_global::<ServerOutput, _>(
+        3,
+        ways::Filter::new(move |(newo, _): (ways::Main<ServerOutput>, _), _, _| {
             let destructor_called_resource = destructor_called_global.clone();
             newo.quick_assign(|_, _, _| {});
             newo.assign_destructor(ways::Filter::new(move |_: ways::Resource<_>, _, _| {
                 *destructor_called_resource.lock().unwrap() = true;
             }));
-        });
+        }),
+    );
 
     let mut client = TestClient::new(&server.socket_name);
     let manager = wayc::GlobalManager::new(&client.display_proxy);
@@ -46,14 +47,15 @@ fn resource_destructor_cleanup() {
     let destructor_called_global = destructor_called.clone();
 
     let mut server = TestServer::new();
-    server
-        .display
-        .create_global::<ServerOutput, _>(3, move |newo, _, _| {
+    server.display.create_global::<ServerOutput, _>(
+        3,
+        ways::Filter::new(move |(newo, _): (ways::Main<ServerOutput>, _), _, _| {
             let destructor_called_resource = destructor_called_global.clone();
             newo.assign_destructor(ways::Filter::new(move |_: ways::Resource<_>, _, _| {
                 *destructor_called_resource.lock().unwrap() = true;
             }));
-        });
+        }),
+    );
 
     let mut client = TestClient::new(&server.socket_name);
     let manager = wayc::GlobalManager::new(&client.display_proxy);
@@ -78,15 +80,16 @@ fn client_destructor_cleanup() {
     let destructor_called_global = destructor_called.clone();
 
     let mut server = TestServer::new();
-    server
-        .display
-        .create_global::<ServerOutput, _>(3, move |output, _, _| {
+    server.display.create_global::<ServerOutput, _>(
+        3,
+        ways::Filter::new(move |(output, _): (ways::Main<ServerOutput>, _), _, _| {
             let destructor_called_resource = destructor_called_global.clone();
             let client = output.as_ref().client().unwrap();
             client.add_destructor(ways::Filter::new(move |_, _, _| {
                 *destructor_called_resource.lock().unwrap() = true;
             }));
-        });
+        }),
+    );
 
     let mut client = TestClient::new(&server.socket_name);
     let manager = wayc::GlobalManager::new(&client.display_proxy);