+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [client] The `attach()` method of `Proxy<_>` and the `detach()` method of `Attached<_>` now take `self` by
   reference instead of by value, allowing the creation of new attached/detached proxy handles without
   requiring ownership
+- [server] The globals-creation APIs now take a `Filter`, as the rest of the callbacks of the crate.
 
 ## 0.24.1 -- 2019-12-13